Rename src/{pathpool => entrypool}.rs
